Implement container for todo modal
